Bugfix locking of button readout
The frontpanel button read out had an bug, so that the change of the matrix row locked up because of the debounce on the release of the buttons.

So there is now a variable which locks the frontpanel buttons after a registered button release and a routine ensuring further switching of the matrix by comparing the debounce timeout against the lock variable, hindering the switching.